requirements.txt
fastapi
uvicorn
transformers
torch
motor
langchain
openai
pandas
numpy
scikit-learn
pymongo
python-multipart
motor[srv]
redis
PyJWT